[cffLib] Move desubroutinize from subset module to CFFFontSet
This way we can use it from both subset and merge modules.
